move teams to file
