+++ wire-ios-sync-engine/Tests/Source/SessionManager/SessionManagerTests.swift
@@ -457,137 +457,6 @@ extension IntegrationTest {
 
 // MARK: - Account Deletion
 
-class SessionManagerTests_Teams: IntegrationTest {
-
-    override func setUp() {
-        super.setUp()
-        createSelfUserAndConversation()
-    }
-
-    func testThatItUpdatesAccountAfterLoginWithTeamName() {
-        // given
-        let teamName = "Wire"
-        let image = MockAsset(in: mockTransportSession.managedObjectContext, forID: selfUser.previewProfileAssetIdentifier!)
-        self.mockTransportSession.performRemoteChanges { session in
-            let team = session.insertTeam(withName: teamName, isBound: true, users: [self.selfUser])
-            team.creator = self.selfUser
-        }
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-
-        // when
-        XCTAssert(login())
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-
-        _ = MockAsset(in: mockTransportSession.managedObjectContext, forID: selfUser.previewProfileAssetIdentifier!)
-
-        // then
-        guard let sharedContainer = Bundle.main.appGroupIdentifier.map(FileManager.sharedContainerDirectory) else { return XCTFail() }
-        let manager = AccountManager(sharedDirectory: sharedContainer)
-        guard let account = manager.accounts.first, manager.accounts.count == 1 else { XCTFail("Should have one account"); return }
-        XCTAssertEqual(account.userIdentifier.transportString(), self.selfUser.identifier)
-        XCTAssertEqual(account.teamName, teamName)
-        XCTAssertEqual(account.imageData, image?.data)
-        XCTAssertNil(account.teamImageData)
-        XCTAssertEqual(account.loginCredentials, selfUser.loginCredentials)
-    }
-
-    func testThatItUpdatesAccountWithUserDetailsAfterLogin() {
-        // when
-        XCTAssert(login())
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-
-        // then
-        guard let sharedContainer = Bundle.main.appGroupIdentifier.map(FileManager.sharedContainerDirectory) else { return XCTFail() }
-        let manager = AccountManager(sharedDirectory: sharedContainer)
-        guard let account = manager.accounts.first, manager.accounts.count == 1 else { XCTFail("Should have one account"); return }
-        XCTAssertEqual(account.userIdentifier.transportString(), self.selfUser.identifier)
-        XCTAssertNil(account.teamName)
-        XCTAssertEqual(account.userName, self.selfUser.name)
-        let image = MockAsset(in: mockTransportSession.managedObjectContext, forID: selfUser.previewProfileAssetIdentifier!)
-
-        XCTAssertEqual(account.imageData, image?.data)
-    }
-
-    func testThatItUpdatesAccountWithUserDetailsAfterLoginIntoExistingAccount() {
-        // given
-        XCTAssert(login())
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-
-        // when
-        sessionManager?.logoutCurrentSession()
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-        closePushChannelAndWaitUntilClosed()
-        XCTAssert(login())
-
-        // then
-        guard let sharedContainer = Bundle.main.appGroupIdentifier.map(FileManager.sharedContainerDirectory) else { return XCTFail() }
-        let manager = AccountManager(sharedDirectory: sharedContainer)
-        guard let account = manager.accounts.first, manager.accounts.count == 1 else { XCTFail("Should have one account"); return }
-        XCTAssertEqual(account.userIdentifier.transportString(), self.selfUser.identifier)
-        XCTAssertNil(account.teamName)
-        XCTAssertEqual(account.userName, self.selfUser.name)
-        let image = MockAsset(in: mockTransportSession.managedObjectContext, forID: selfUser.previewProfileAssetIdentifier!)
-
-        XCTAssertEqual(account.imageData, image?.data)
-    }
-
-    func testThatItUpdatesAccountAfterUserNameChange() {
-        // when
-        XCTAssert(login())
-
-        let newName = "BOB"
-        self.mockTransportSession.performRemoteChanges { _ in
-            self.selfUser.name = newName
-        }
-        XCTAssertTrue(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-
-        // then
-        guard let sharedContainer = Bundle.main.appGroupIdentifier.map(FileManager.sharedContainerDirectory) else { return XCTFail() }
-        let manager = AccountManager(sharedDirectory: sharedContainer)
-        guard let account = manager.accounts.first, manager.accounts.count == 1 else { XCTFail("Should have one account"); return }
-        XCTAssertEqual(account.userIdentifier.transportString(), self.selfUser.identifier)
-        XCTAssertNil(account.teamName)
-        XCTAssertEqual(account.userName, selfUser.name)
-    }
-
-    func testThatItSendsAuthenticationErrorWhenAccountLimitIsReached() throws {
-        // given
-        let account1 = Account(userName: "Account 1", userIdentifier: UUID.create())
-        let account2 = Account(userName: "Account 2", userIdentifier: UUID.create())
-        let account3 = Account(userName: "Account 3", userIdentifier: UUID.create())
-
-        sessionManager?.accountManager.addOrUpdate(account1)
-        sessionManager?.accountManager.addOrUpdate(account2)
-        sessionManager?.accountManager.addOrUpdate(account3)
-
-        // when
-        XCTAssert(login(ignoreAuthenticationFailures: true))
-
-        // then
-        guard
-            let delegate = mockLoginDelegete,
-            let error = delegate.currentError
-        else {
-            return XCTFail()
-        }
-
-        XCTAssertTrue(delegate.didCallAuthenticationDidFail)
-        XCTAssertEqual(error, NSError(code: .accountLimitReached, userInfo: nil))
-    }
-
-    func testThatItChecksAccountsForExistingAccount() {
-        // given
-        let account1 = Account(userName: "Account 1", userIdentifier: UUID.create())
-        let account2 = Account(userName: "Account 2", userIdentifier: UUID.create())
-
-        sessionManager?.accountManager.addOrUpdate(account1)
-
-        // then
-        XCTAssertTrue(sessionManager!.session(session: self.unauthenticatedSession!, isExistingAccount: account1))
-        XCTAssertFalse(sessionManager!.session(session: self.unauthenticatedSession!, isExistingAccount: account2))
-    }
-}
-
 final class SessionManagerTests_MultiUserSession: IntegrationTest {
 
     override func setUp() {
